% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00area_calc.R
\name{area_calc}
\alias{area_calc}
\title{area_calc}
\usage{
area_calc(data_in, faixa_dist = c(0.5, 10, 20, 22), faixa_lado = c(1.5, 2,
  10, 21.5, 22), area_epsg = 3395)
}
\arguments{
\item{data_in}{Data.frame with bearings and distances. See \code{data(br319)}
for example of required format.}

\item{faixa_dist}{Distances to calculate area.}

\item{faixa_lado}{Distances to calculate area, 
seperately to the left and right of the central line}

\item{area_epsg}{EPSG for projection. This gives coordinate reference system used 
when calculating area.}
}
\value{
List of data and spatial (lines and polygons).
}
\description{
\code{area_calc} Generates lines and polygons for area calculation of nonlinear plots.
}
\examples{
\dontrun{
# Example with required data format
data(br319)
dados_in <- br319

# Generate lines and polygons for specified widths
list_res <- parcelareadev::area_calc(
data_in = dados_in, 
faixa_dist = c(0.5, 1, 3,12, 10, 20, 21, 22),
faixa_lado = c(0.5, 1, 3,12, 10, 20, 21, 22),
area_epsg = 3395)

# Calculate area and export results
df.resumo <- parcelareadev::area_results(results_list = list_res, 
                       make_shape = FALSE)
}
}

