% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11area_results.R
\name{area_results}
\alias{area_results}
\title{area_results}
\usage{
area_results(results_list = list_res, make_shape = FALSE)
}
\arguments{
\item{results_list}{List of results from \code{area_calc}.}

\item{make_shape}{Logical, should shapes be exported.}
}
\value{
Exports .csv of calulated area values and .pdfs with figures for checking.
}
\description{
\code{area_results} Calculates area and exports results in user friendly formats.
}
\examples{
\dontrun{
# Example with required data format
data(br319)
dados_in <- br319

# Generate lines and polygons for specified widths
list_res <- parcelareadev::area_calc(
data_in = dados_in, 
faixa_dist = c(0.5, 1, 3,12, 10, 20, 21, 22),
faixa_lado = c(0.5, 1, 3,12, 10, 20, 21, 22),
area_epsg = 3395)

# Calculate area and export results
df.resumo <- parcelareadev::area_results(results_list = list_res,
                                        make_shape = FALSE)
                                        
# Now shift central line to geographic coordinates 
# and export as shapefiles and KML for visualization and checking
data("br319coords")
pc <- br319coords
parcelareadev::exp_results(results_list = list_res, 
                           pcoords = pc, exp_KML = TRUE)                                       
}
}
