% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{br319}
\alias{br319}
\title{Data from 5 PPBIO plots along the BR319 in Brazil.}
\format{A data frame with 138 rows and 6 variables:
\describe{
  \item{aid}{unique rowid, for data checking.}
  \item{plot_id}{ID code, unique for each plot.}
  \item{azimute}{bearing, used to calculate coordinates.}
  \item{segmento}{Distance (in meters) of each segment.}
  \item{remove}{Flag. Should segment be removed from analysis}
  \item{seg_id}{ID for segments. Unique within plot, for ordering.}
}}
\source{
\url{https://ppbio.inpa.gov.br/sitios/br319/}
}
\usage{
br319
}
\description{
A dataset containing the azimuth and length of segments from 5 plots:
"M01_TN_0500", "M02_TN_0500","M02_TN_2500","M03_TN_1500", "M03_TN_2500"
}
\keyword{datasets}

