% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{br319res}
\alias{br319res}
\title{Calculated areas from 9 PPBIO plots along the BR319 in Brazil.}
\format{A data frame with 864 rows and 8 variables:
\describe{
  \item{.id}{ID code, unique for each plot.}
  \item{variable}{Indicates which segments were used (4 types).}
  \item{lado}{Indicates which side was used for area calculations.}
  \item{faixa_id}{ID used in width calculation.}
  \item{area_m2}{Calculated area in m^2.}
  \item{seg_count}{How many segments were used in the area calculation.}
  \item{faixa_tipo}{ID for grouping calculation types.}
  \item{faixa_dist}{Indicates buffer width used in area calculations.}
}}
\usage{
br319res
}
\description{
A dataset containing the calculated area (m2) from 9 nonlinear plots:
M01_TN_0500, M01_TN_3500, M02_TN_1500, M06_TS_0500, M02_TN_0500, M02_TN_2500,
M02_TS_0500, M03_TN_2500, M09_TN_1500 .
}
\keyword{datasets}

