% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09fig_check_area.R
\name{check_area}
\alias{check_area}
\title{check_area}
\usage{
check_area(x, c_epsg = 3395)
}
\arguments{
\item{x}{List with data, spatial lines and polygons generated by \code{area_calc}.}

\item{c_epsg}{EPSG code for projected buffers.}
}
\value{
Figure with polygons of widths.
}
\description{
Should not be called directly by users.
\code{check_area} plots buffer (spatial polygons) for widths around central line.
}
\examples{
\dontrun{
}
}
