% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10fig_check_lado.R
\name{check_lado}
\alias{check_lado}
\title{check_lado}
\usage{
check_lado(x, cl_epsg = 3395)
}
\arguments{
\item{x}{List with data, spatial lines and polygons generated by \code{area_calc}.}

\item{cl_epsg}{EPSG code for projected buffers.}
}
\value{
Figure with left and right buffers.
}
\description{
Should not be called directly by users.
\code{check_lado} plots left and right buffer polygons.
}
\examples{
\dontrun{
}
}

