% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08fig_check_linha.R
\name{check_linha}
\alias{check_linha}
\title{check_linha}
\usage{
check_linha(x)
}
\arguments{
\item{x}{List with data, spatial lines and polygons generated by \code{area_calc}.}
}
\value{
Figure with center lines.
}
\description{
Should not be called directly by users.
\code{check_linha} plots central line and distances.
}
\examples{
\dontrun{
}
}

