% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01checkAngle.R
\name{checkAngle}
\alias{checkAngle}
\title{checkAngle}
\usage{
checkAngle(x)
}
\arguments{
\item{x}{Data.frame with bearings and distances, as per \code{data(br319)} .}
}
\value{
Identifies which angles need to be excluded
}
\description{
Should not be called directly by users.
\code{checkAngle} identifies which segments are seperated by less than 70 degrees.
}
\examples{
\dontrun{
}
}
