% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctAng.R
\name{correctAng}
\alias{correctAng}
\title{correctAng}
\usage{
correctAng(x)
}
\arguments{
\item{x}{Data.frame}
}
\value{
Angle corrected from azimuth calculation
}
\description{
Should not be called directly by users.
\code{correctAng} corrects bearings, prior to procesing by \code{checkAngle}.
}
\examples{
\dontrun{
}
}
