% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05dist2gcS.R
\name{dist2gcS}
\alias{dist2gcS}
\title{dist2gcs}
\usage{
dist2gcS(p1, p2, p3, r = 6378137)
}
\arguments{
\item{p1}{Start of great circle path. longitude/latitude of point(s).}

\item{p2}{End of great circle path. As above.}

\item{p3}{Point away from the great cricle path. As for p2.}

\item{r}{radius of the earth; default = 6378137.}
}
\value{
Signed distance between point and line i.e. cross track distance.
}
\description{
Should not be called directly by users.
Copied and adapted from function "dist2gc" in geosphere package by Robert Hijmans.
\code{dist2gcs} calculates signed distance between point and a line.
}
\details{
Just took out "abs" from original code in \code{geosphere::dist2gc}. 
So returns signed values to indicate left or right, as used by \code{makelado}.
}
\examples{
\dontrun{
dist2gcs(c(0,0),c(90,90),c(80,80)
dist2gcs(c(0,0),c(90,90),c(-80,80))
}
}
