% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04doLine.R
\name{doLine}
\alias{doLine}
\title{doLine}
\usage{
doLine(dfseg, line_epsg = 3395)
}
\arguments{
\item{dfseg}{Data.frame of coordinates from segments generated by \code{makeSeg}.}

\item{line_epsg}{EPSG code to use.}
}
\value{
Produces list with two SpatialLinesDataFrames, one with geographic (epsg = 4326)
and the other projected (epsg:3395) coordinate reference systems.
}
\description{
Should not be called directly by users.
\code{doline} creates central line from coordinates generated by \code{makeSeg}.
}
\examples{
\dontrun{
}
}

