% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12exp_results.R
\name{exp_results}
\alias{exp_results}
\title{exp_results}
\usage{
exp_results(results_list = list_res, pcoords = pcoords, exp_KML = TRUE)
}
\arguments{
\item{results_list}{List of results from \code{area_calc}.}

\item{pcoords}{Data.frame, with geographic coordinates of plot start points.}

\item{exp_KML}{Logical. Export as KML.}
}
\value{
Exports shapefiles and kml of central line shifted to geographic coordinates.
}
\description{
\code{exp_results} Exports central line shifted to geographic coordinates.
}
\examples{
\dontrun{
# Example with required data format
data(br319)
dados_in <- br319

# Generate lines and polygons for specified widths
list_res <- parcelareadev::area_calc(
data_in = dados_in, 
faixa_dist = c(0.5, 1, 3,12, 10, 20, 21, 22),
faixa_lado = c(0.5, 1, 3,12, 10, 20, 21, 22),
area_epsg = 3395)

# Calculate area and export results
df.resumo <- parcelareadev::area_results(results_list = list_res,
                                        make_shape = FALSE)
                                        
# Now shift central line to geographic coordinates 
# and export as shapefiles and KML for visualization and checking
data("br319coords")
pc <- br319coords
parcelareadev::exp_results(results_list = list_res, 
                           pcoords = pc, exp_KML = TRUE)                                         
}
}
