% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expshape.R
\name{expshape}
\alias{expshape}
\title{expshape}
\usage{
expshape(x, c_epsg = 3395)
}
\arguments{
\item{x}{List with data, spatial lines and polygons generated by \code{area_calc}.}

\item{c_epsg}{EPSG code for projected buffers.}
}
\value{
Exports shapefiles.
}
\description{
Should not be called directly by users.
\code{expshape} exports shapefiles of lines and buffers.
}
\examples{
\dontrun{
}
}

