% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expshiftKML.R
\name{expshiftKML}
\alias{expshiftKML}
\title{expshiftkML}
\usage{
expshiftKML(res_list = res_list, pcoords = pcoords, c_epsg = 3395)
}
\arguments{
\item{res_list}{List with data, spatial lines and polygons generated by \code{area_calc}.}

\item{pcoords}{Data.frame with geographic coordinates of plot start point.}

\item{c_epsg}{EPSG code for projected buffers.}
}
\value{
Exports cental line shifted to geographic coordinates 
as kml for visualization.
}
\description{
Should not be called directly by users.
\code{expshiftKML} Moves central line to geographic coordinates.
}
\examples{
\dontrun{
}
}
