% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02getCoords.R
\name{getCoords}
\alias{getCoords}
\title{getCoords}
\usage{
getCoords(data)
}
\arguments{
\item{data}{Data.frame with bearings and distances, as per \code{data(br319)} .}
}
\value{
Coordinates from bearings and distance between points
}
\description{
Should not be called directly by users.
\code{getCoords} generates coordinates from bearing and distance.
}
\examples{
\dontrun{
}
}
