% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06makeLado.R
\name{makelado}
\alias{makelado}
\title{makelado}
\usage{
makelado(list.lines, faixa_lado = c(0.5, 1, 1.5, 2, 10, 20),
  lado_epsg = 3395)
}
\arguments{
\item{list.lines}{List with central lines (class SpatialLines) generated by \code{doline}.}

\item{faixa_lado}{Vector of plot survey widths.}

\item{lado_epsg}{EPSG code to use.}
}
\value{
List with SpatialPolygons (buffers) to left and right of central line.
}
\description{
Should not be called directly by users.
\code{makelado} creates buffers to the left and right of central line.
}
\examples{
\dontrun{
}
}

