% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03makeSeg.R
\name{makeSeg}
\alias{makeSeg}
\title{makeSeg}
\usage{
makeSeg(df.coords)
}
\arguments{
\item{df.coords}{Data.frame generated by \code{getCoords}}
}
\value{
data.frame with segments
}
\description{
Should not be called directly by users.
\code{makeSeg} generates segments from coordinates generated by \code{getCoords}.
}
\examples{
\dontrun{
}
}
