% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07out_resumo.R
\name{out_resumo}
\alias{out_resumo}
\title{out_resumo}
\usage{
out_resumo(list.poly)
}
\arguments{
\item{list.poly}{List of results from \code{area_calc}}
}
\value{
Data.frame with calculated areas.
}
\description{
Should not be called directly by users.
\code{out_resumo} calculates area and formats results for .csv export.
}
\examples{
\dontrun{
}
}
