% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05dist2gcS.R
\name{dist2gcS}
\alias{dist2gcS}
\title{dist2gcs}
\usage{
dist2gcS(p1, p2, p3, r = 6378137)
}
\arguments{
\item{p1}{}

\item{p2}{}

\item{p3}{}

\item{r}{}
}
\value{
Signed distance between point and line i.e. cross track distance.
}
\description{
Should not be called directly by users.
Copied and adapted from function "dist2gc" in geosphere package by Robert Hijmans.
\code{dist2gcs} calculates signed distance between point and a line.
}
\details{
Just took out "abs" from original code in \code{geosphere::dist2gc}. 
So returns signed values to indicate left or right, as used by \code{makelado}.
}
\examples{
\dontrun{
}
}

